:truck: move mdx components to components dir
sounds obvious in retrospect
